use macdrive Go library instead of webview
