+++ client/cmd/dexc-desktop/go.mod
@@ -8,6 +8,7 @@ require (
 	decred.org/dcrdex v0.6.1
 	fyne.io/systray v1.10.1-0.20230403195833-7dc3c09283d6
 	github.com/gen2brain/beeep v0.0.0-20220909211152-5a9ec94374f6
+	github.com/progrium/macdriver v0.2.0
 	github.com/webview/webview v0.0.0-20230415172654-8387ff8945fc
 )
 
@@ -146,6 +147,7 @@ require (
 	github.com/mitchellh/pointerstructure v1.2.0 // indirect
 	github.com/nu7hatch/gouuid v0.0.0-20131221200532-179d4d0c4d8d // indirect
 	github.com/olekukonko/tablewriter v0.0.5 // indirect
+	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/prometheus/client_golang v1.14.0 // indirect
 	github.com/prometheus/client_model v0.3.0 // indirect
